xfd,morebits: Recognise DST in date formatting
Wikis other than English Wikipedia often don't use UTC, but rather a
country-specific timezone, and one that may also observe DST. This
means that the code needed to be updated twice a year or was wrong
for an hour every day for half the year resulting in e.g. XFD messages
being left on the wrong page.

Fix this by supporting a timezone name and using the Intl API to
resolve it to a minute offset. The browser internally will use its
tzdatabase to know what timezone offset the name translates to at any
given date and time, including to correctly shift e.g. from +120 to
+60 on 30 October 02:00 AM etc.